Improved VSCode dev experience

+++ build/watchWebClient.ps1
@@ -1,3 +1,10 @@
+[CmdletBinding()]
+param (
+  [Parameter()]
+  [switch]
+  $DisableVueDevServer
+)
+
 $originalLocation = Get-Location
 $projectRoot = "$PSScriptRoot/../"
 
@@ -5,13 +12,20 @@ try {
   Set-Location -Path $projectRoot
   . ./build/buildSettings.ps1
 
-  Remove-Item -Path "$webProjectFolder/wwwroot/*" -Recurse -ErrorAction SilentlyContinue
-  Copy-Item -Path "$webClientProjectFolder/public/*" -Destination "$webProjectFolder/wwwroot" -Recurse
+  if (-not $DisableVueDevServer) {
+    Remove-Item -Path "$webProjectFolder/wwwroot/*" -Recurse -ErrorAction SilentlyContinue
+    Copy-Item -Path "$webClientProjectFolder/public/*" -Destination "$webProjectFolder/wwwroot" -Recurse
 
-  Set-Location -Path $webClientProjectFolder
+    Set-Location -Path $webClientProjectFolder
+    npm install --no-audit
+    npm run dev
 
-  npm install --no-audit
-  npm run dev
+  } else {
+    Set-Location -Path $webClientProjectFolder
+    npm install --no-audit
+    npm run watch
+  }
+
 
 } finally {
   Set-Location $originalLocation
